19/7/2019 4rd

+++ database/migrations/2019_06_28_060420_create_accounts_table.php
@@ -20,7 +20,7 @@ class CreateAccountsTable extends Migration
             $table->integer('balance');
             $table->string('branch_name')->nullable();
             $table->string('branch_code')->nullable();
-            $table->integer('account_number')->nullable();
+            $table->bigInteger('account_number')->nullable();
             $table->timestamps();
         });
     }
